+++ packages/gi-sdk/src/utils/index.ts
@@ -1,1 +1,2 @@
+export { createDownload } from './download';
 export { Shortcuts, useShortcuts } from './shortcuts';