+++ packages/gi-site/src/pages/Workspace/Projects.tsx
@@ -101,100 +101,106 @@ const ProjectList: React.FunctionComponent<ProjectListProps> = props => {
     );
   }
   return (
-    <>
-      <Row gutter={[16, 16]} style={{ paddingRight: '24px' }}>
-        {[...lists]
-          .sort((p1, p2) => p2.gmtCreate - p1.gmtCreate)
-          .map(item => {
-            const { id, name, gmtCreate, recycleTime, cover } = item;
-            let expiredStr;
-            if (recycleTime) {
-              const expiredDate = new Date(recycleTime + 604800000);
-              expiredStr = `${expiredDate.toLocaleDateString()} ${expiredDate.toLocaleTimeString()}`;
-            }
-            const time = utils.time(gmtCreate);
-            const Cover = (
+    <Row gutter={[16, 16]} style={{ paddingRight: '24px' }}>
+      {[...lists]
+        .sort((p1, p2) => p2.gmtCreate - p1.gmtCreate)
+        .map(item => {
+          const { id, name, gmtCreate, recycleTime, cover } = item;
+          let expiredStr;
+          if (recycleTime) {
+            const expiredDate = new Date(recycleTime + 604800000);
+            expiredStr = `${expiredDate.toLocaleDateString()} ${expiredDate.toLocaleTimeString()}`;
+          }
+          const time = utils.time(gmtCreate);
+          const Cover = (
+            <div
+              style={{
+                width: '100%',
+                height: 250,
+                display: 'flex',
+                justifyContent: 'center',
+                alignItems: 'center',
+                padding: 5,
+                overflow: 'hidden',
+              }}
+            >
               <img
                 src={cover || `${window['GI_PUBLIC_PATH']}image/empty_workbook.png`}
-                style={{ cursor: 'pointer', width: '100%' }}
+                style={{ cursor: 'pointer', height: '100%' }}
                 onClick={() => {
                   history.push(`/workspace/${id}?nav=style`);
                 }}
               />
-            );
+            </div>
+          );
 
-            const expiredInfo = expiredStr && (
-              <span style={{ color: 'red' }}>
-                {$i18n.get({ id: 'gi-site.pages.Workspace.Projects.DataSourceDeleted', dm: '数据源已删除' })}
+          const expiredInfo = expiredStr && (
+            <span style={{ color: 'red' }}>
+              {$i18n.get({ id: 'gi-site.pages.Workspace.Projects.DataSourceDeleted', dm: '数据源已删除' })}
 
-                <Tooltip
-                  title={$i18n.get(
-                    {
-                      id: 'gi-site.pages.Workspace.Projects.TheRelevantDatasetHasBeen',
-                      dm: '相关数据集已删除，该工作簿将于 {expiredStr} 过期自动销毁。若需恢复，请在「数据集-回收站」恢复相关数据',
-                    },
-                    { expiredStr: expiredStr },
-                  )}
-                >
-                  <QuestionCircleOutlined style={{ cursor: 'pointer' }} />
-                </Tooltip>
-              </span>
-            );
+              <Tooltip
+                title={$i18n.get(
+                  {
+                    id: 'gi-site.pages.Workspace.Projects.TheRelevantDatasetHasBeen',
+                    dm: '相关数据集已删除，该工作簿将于 {expiredStr} 过期自动销毁。若需恢复，请在「数据集-回收站」恢复相关数据',
+                  },
+                  { expiredStr: expiredStr },
+                )}
+              >
+                <QuestionCircleOutlined style={{ cursor: 'pointer' }} />
+              </Tooltip>
+            </span>
+          );
 
-            const description = (
-              <>
-                {time} {expiredInfo}
-              </>
-            );
+          const description = (
+            <>
+              {time} {expiredInfo}
+            </>
+          );
 
-            return (
-              <Col key={id} xs={24} sm={24} md={12} lg={12} xl={8} xxl={6}>
-                <Card cover={Cover}>
-                  <div style={{ position: 'relative' }}>
-                    <Meta
-                      title={
-                        <Input
-                          defaultValue={name}
-                          bordered={false}
-                          onBlur={e => {
-                            const value = e.target.value;
-                            if (id && !isNil(value) && value !== name) {
-                              ProjectService.updateById(id, { name: value });
-                            }
-                          }}
-                        />
-                      }
-                      description={description}
-                    />
-
-                    <div style={{ position: 'absolute', bottom: '0px', right: '0px' }}>
-                      <Popconfirm
-                        title={$i18n.get({
-                          id: 'gi-site.pages.Workspace.Projects.DeleteTheProject',
-                          dm: '是否删除该项目?',
-                        })}
-                        onConfirm={e => {
-                          e!.preventDefault();
-                          confirm(id);
+          return (
+            <Col key={id} xs={24} sm={24} md={12} lg={12} xl={8} xxl={6}>
+              <Card cover={Cover}>
+                <div style={{ position: 'relative' }}>
+                  <Meta
+                    title={
+                      <Input
+                        defaultValue={name}
+                        bordered={false}
+                        onBlur={e => {
+                          const value = e.target.value;
+                          if (id && !isNil(value) && value !== name) {
+                            ProjectService.updateById(id, { name: value });
+                          }
                         }}
-                        placement="rightBottom"
-                        okText="Yes"
-                        cancelText="No"
-                      >
-                        <Button
-                          type="text"
-                          size="small"
-                          icon={<DeleteOutlined className="more icon-buuton" />}
-                        ></Button>
-                      </Popconfirm>
-                    </div>
+                      />
+                    }
+                    description={description}
+                  />
+
+                  <div style={{ position: 'absolute', bottom: '0px', right: '0px' }}>
+                    <Popconfirm
+                      title={$i18n.get({
+                        id: 'gi-site.pages.Workspace.Projects.DeleteTheProject',
+                        dm: '是否删除该项目?',
+                      })}
+                      onConfirm={e => {
+                        e!.preventDefault();
+                        confirm(id);
+                      }}
+                      placement="rightBottom"
+                      okText="Yes"
+                      cancelText="No"
+                    >
+                      <Button type="text" size="small" icon={<DeleteOutlined className="more icon-buuton" />}></Button>
+                    </Popconfirm>
                   </div>
-                </Card>
-              </Col>
-            );
-          })}
-      </Row>
-    </>
+                </div>
+              </Card>
+            </Col>
+          );
+        })}
+    </Row>
   );
 };