+++ packages/gi-assets-galaxybase/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @antv/gi-assets-galaxybase
 
+## 1.2.5
+
+### Patch Changes
+
+- publish in 20230728
+- Updated dependencies
+  - @antv/gi-sdk@2.4.5
+
+## 1.2.4
+
+### Patch Changes
+
+- publish in 0720 pm
+- Updated dependencies
+  - @antv/gi-sdk@2.4.4
+
+## 1.2.3
+
+### Patch Changes
+
+- publish in 2023-07-20
+- Updated dependencies
+  - @antv/gi-sdk@2.4.3
+
 ## 1.2.2
 
 ### Patch Changes