+++ packages/gi-httpservice/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @antv/gi-httpservice
 
+## 1.2.5
+
+### Patch Changes
+
+- publish in 20230728
+
+## 1.2.4
+
+### Patch Changes
+
+- publish in 0720 pm
+
+## 1.2.3
+
+### Patch Changes
+
+- publish in 2023-07-20
+
 ## 1.2.2
 
 ### Patch Changes