+++ packages/gi-sdk-app/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @antv/gi-sdk-app
 
+## 1.0.4
+
+### Patch Changes
+
+- publish in 20230728
+
+## 1.0.3
+
+### Patch Changes
+
+- publish in 0720 pm
+
+## 1.0.2
+
+### Patch Changes
+
+- publish in 2023-07-20
+
 ## 1.0.1
 
 ### Patch Changes