+++ packages/gi-theme-antd/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @antv/gi-theme-antd
 
+## 0.6.5
+
+### Patch Changes
+
+- publish in 20230728
+
+## 0.6.4
+
+### Patch Changes
+
+- publish in 0720 pm
+
+## 0.6.3
+
+### Patch Changes
+
+- publish in 2023-07-20
+
 ## 0.6.2
 
 ### Patch Changes